remade game structure and split context
