+++ client/src/App.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
 import "./App.css";
-import { Provider } from "./services/Context";
+import { Provider } from "./services/GameContext";
 
 function App() {
   return (
