--- client/src/components/game/gameBoard/room/Html.tsx
@@ -1,17 +1,0 @@
-import "../GameBoard.css";
-
-function Html() {
-  return (
-    <>
-      <div className="gameboard-container">
-        <img
-          className="game-img"
-          src="./src/assets/images/html-room.png"
-          alt="plateau de jeu"
-        />
-      </div>
-    </>
-  );
-}
-
-export default Html;
