+++ client/src/components/game/gameCommands/GameCommands.tsx
@@ -1,4 +1,4 @@
-import { Context } from "../../../services/Context";
+import { GameContext } from "../../../services/GameContext";
 import "./GameCommands.css";
 import { useContext } from "react";
 import Prompt from "./minigame/prompt/Prompt";
@@ -6,7 +6,7 @@ import Quizz from "./minigame/quizz/Quizz";
 
 function GameCommands() {
   //Importation du contexte
-  const context = useContext(Context);
+  const context = useContext(GameContext);
 
   if (!context) {
     return <div>Error: Context is not available</div>;
