+++ client/src/pages/game/GameBoard.tsx
@@ -1,5 +1,5 @@
-import "./Game.css";
-import GameBoard from "../../components/game/gameBoard/GameBoard";
+import "./GameBoard.css";
+import GameDisplay from "../../components/game/gameDisplay/GameDisplay";
 import GameCommands from "../../components/game/gameCommands/GameCommands";
 import GameInstructions from "../../components/game/gameInstructions/GameInstructions";
 import GameProfil from "../../components/game/gameProfil/GameProfil";
@@ -8,7 +8,7 @@ function Game() {
   return (
     <>
       <div className="game-page">
-        <GameBoard />
+        <GameDisplay />
         <GameProfil />
         <GameInstructions />
         <GameCommands />
