+++ client/src/services/GameContext.tsx
@@ -5,10 +5,6 @@ import { useEffect } from "react";
 interface ContextValue {
   challenge: ChallengeProps[];
   setChallenge: React.Dispatch<React.SetStateAction<ChallengeProps[]>>;
-  account: AccountProps[];
-  setAccount: React.Dispatch<React.SetStateAction<AccountProps[]>>;
-  progress: ProgressProps[];
-  setProgress: React.Dispatch<React.SetStateAction<ProgressProps[]>>;
   currentIndex: number;
   setCurrentIndex: React.Dispatch<React.SetStateAction<number>>;
   currentType: number;
@@ -50,30 +46,15 @@ interface ChallengeProps {
   room_id: number;
 }
 
-interface AccountProps {
-  id: number;
-  username: string;
-  email: string;
-  password: string;
-}
-
-interface ProgressProps {
-  level: number;
-  user_id: number;
-  room_id: number;
-  challenge_id: number;
-}
-
 interface ProviderProps {
   children: ReactNode;
 }
 
-export const Context = createContext<ContextValue | null>(null);
+export const GameContext = createContext<ContextValue | null>(null);
 
 export const Provider = ({ children }: ProviderProps) => {
+
   const [challenge, setChallenge] = useState([] as ChallengeProps[]);
-  const [account, setAccount] = useState([] as AccountProps[]);
-  const [progress, setProgress] = useState([] as ProgressProps[]);
   const [currentIndex, setCurrentIndex] = useState(0);
   const [currentType, setCurrentType] = useState(0);
   const [isButtonEnabled, setIsButtonEnabled] = useState(false);
@@ -96,34 +77,12 @@ export const Provider = ({ children }: ProviderProps) => {
   }, []);
 
   //----------------------------------------------------------
-  // FETCH DE LA TABLE ACCOUNT
-  useEffect(() => {
-    fetch(`${import.meta.env.VITE_API_URL}/api/account`)
-      .then((response) => response.json())
-      .then((data: AccountProps[]) => {
-        setAccount(data);
-      });
-  }, []);
-
-  //----------------------------------------------------------
-  // FETCH DE LA TABLE PROGRESS
-  useEffect(() => {
-    fetch(`${import.meta.env.VITE_API_URL}/api/progress`)
-      .then((response) => response.json())
-      .then((data: ProgressProps[]) => {
-        setProgress(data);
-      });
-  }, []);
 
   return (
-    <Context.Provider
+    <GameContext.Provider
       value={{
         challenge,
         setChallenge,
-        account,
-        setAccount,
-        progress,
-        setProgress,
         currentIndex,
         setCurrentIndex,
         currentType,
@@ -147,14 +106,14 @@ export const Provider = ({ children }: ProviderProps) => {
       }}
     >
       {children}
-    </Context.Provider>
+    </GameContext.Provider>
   );
 };
 
 export const useGameContext = () => {
-  const context = useContext(Context);
-  if (!context) {
+  const gameContext = useContext(GameContext);
+  if (!gameContext) {
     throw new Error("useGameContext must be used within a Provider");
   }
-  return context;
+  return GameContext;
 };
